Create default 16x16 grid and hover effect
Assignment steps 1-3: index.html, js, and css files
linked. The default starting 16x16 grid in constant
640x640px container dimensions created via for loop that calls
a getSquare(). getSquare() creates a div in the DOM
styled by the .css as a flexbox square that will fill
the container as a grid.
Hover effect created as an event listener applied to
all squares that changes background color when called.