Add new grid button and its functionality
Within an event listener for the new grid button, takes
user inputted value (for how many squares per side of new grid)
between 1-100, clears the gridContainer, then uses
getSquare() in a loop to create the appropriate number
of squares and assign the right height/width to each.
To do so, getSquare() changed to return the square (div)
it creates. The hover effect also changed to be contained
in hoverEffect() function so it can be reenabled for
each new grid.

All possible grid squares are contained within gridContainer
by set dimensions and flexbox rules in styles.css.